+++ tests/structured_prediction/predictors/biaffine_dependency_parser_test.py
@@ -19,6 +19,8 @@ class TestBiaffineDependencyParser(AllenNlpTestCase):
         )
         predictor = Predictor.from_archive(archive, "biaffine_dependency_parser")
 
+        print(predictor)
+
         result = predictor.predict_json(inputs)
 
         words = result.get("words")